Step 7: Create functions in HomeController
A. like(index) – a function called “like” that accepts the parameter
“index” (Number). When
called, it should increment the variable “like” variable by one, for the
movie at the given
“index”. (These do not have to save when you refresh the page).
B. dislike(index) – a function called “dislike” that accepts the
parameter “index”. when called, it should increment the variable
“dislike” variable by one, for the movie at the given “index” (Number).
(These do not have to save when you refresh the page).
C. posterClick(index) – a function called “posterClick” that accepts the
parameter “index” (Number). When called, it should increment
posterindex, for the movie at the given “index”, and show the poster at
the updated posterindex index. Once it reaches the last image of the
posters array, it should loop back to the first index (index 0).
D. timeText(minutes) – a function called “timeText” that accepts the
parameter “minutes” (Number). It should convert “minutes” (an integer)
to a String of hours and minutes. For example, “timeText(61)” should
return “1h 1m”, “timeText(145)” should
return “2h 25m”, and “timeText(180)” should return “3h 0m”.