Step 8: Added a Angular click handle
A. .likes to call the like() function in HomeController.js, passing
$index as an argument ($index is a built-in variable that is used by
Angular loops). This should change the displayed value in .likes
B.
.dislikes to call the dislike() function in HomeController.js, passing
$index as an argument ($index is a built-in variable that is used by
Angular loops). This should change the displayed value in .dislikes
C. .poster to call the posterClick() function in HomeController.js,
passing $index as an argument ($index is a built-in variable that is
used by Angular loops). This should change the displayed .poster for the
movie (if there is more than one poster).